拓展图片模块
